requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2
reportlab==4.0.4
groq
requests